+++ src/pages/[nickname]/myProfilePage.tsx
@@ -9,7 +9,7 @@ import {
 import { TemporaryMarkingBar } from "@/entities/marking/ui";
 import { useGetProfile } from "@/entities/profile/api";
 import { ROUTER_PATH } from "@/shared/constants";
-import { useNicknameParams } from "@/shared/lib/profile";
+import { useNicknameParams } from "@/shared/lib";
 import { useAuthStore } from "@/shared/store";
 import { SettingIcon } from "@/shared/ui/icon";
 import { NavigationBar } from "@/shared/ui/navigationbar";
