+++ src/shared/lib/index.ts
@@ -8,3 +8,4 @@ export * from "./error";
 export * from "./snackbar";
 export * from "./scroll";
 export * from "./dropdown";
+export * from "./profile";
